refactor: logical planning for `MERGE`
