+++ crates/sail-delta-lake/src/logical/mod.rs
@@ -1,3 +1,4 @@
+pub mod merge;
 pub mod table_source;
 
 pub use table_source::DeltaTableSource;
