+++ crates/sail-session/Cargo.toml
@@ -27,7 +27,6 @@ sail-physical-optimizer = { path = "../sail-physical-optimizer" }
 sail-plan = { path = "../sail-plan" }
 sail-execution = { path = "../sail-execution" }
 sail-object-store = { path = "../sail-object-store" }
-sail-plan-lakehouse = { path = "../sail-plan-lakehouse" }
 sail-server = { path = "../sail-server" }
 sail-cache = { path = "../sail-cache" }
 sail-telemetry = { path = "../sail-telemetry" }
