+++ crates/sail-session/src/optimizer.rs
@@ -8,13 +8,10 @@ pub fn default_analyzer_rules() -> Vec<Arc<dyn AnalyzerRule + Send + Sync>> {
 
 pub fn default_optimizer_rules() -> Vec<Arc<dyn OptimizerRule + Send + Sync>> {
     let rules = sail_logical_optimizer::default_optimizer_rules();
-    let mut custom = sail_plan_lakehouse::lakehouse_optimizer_rules();
-    custom.extend(
-        rules
-            .into_iter()
-            .filter(|r| r.name() != "push_down_leaf_projections"),
-    );
-    custom
+    rules
+        .into_iter()
+        .filter(|r| r.name() != "push_down_leaf_projections")
+        .collect()
 }
 
 #[cfg(test)]
@@ -22,11 +19,10 @@ mod tests {
     use super::*;
 
     #[test]
-    fn default_rules_start_with_expand_row_level_op() {
+    fn default_rules_skip_leaf_projection_pushdown() {
         let rules = default_optimizer_rules();
-        assert!(
-            rules.first().map(|r| r.name()) == Some("expand_row_level_op"),
-            "expand_row_level_op should run before built-in optimizers"
-        );
+        assert!(!rules
+            .iter()
+            .any(|rule| rule.name() == "push_down_leaf_projections"));
     }
 }
